+++ src_sgct/core/FrameworkInternal.cpp
@@ -137,7 +137,7 @@ namespace viscom {
             auto window = GetEngine()->getWindowPtr(wId);
             window->getFinalFBODimensions(projectorSize.x, projectorSize.y);
             framebuffers_.emplace_back();
-            framebuffers_.back().Resize(projectorSize.x, projectorSize.y);
+            framebuffers_.back().Resize(static_cast<unsigned int>(projectorSize.x), static_cast<unsigned int>(projectorSize.y));
 
             auto vpLocalLowerLeftA = sgct_wrapper::GetProjectionPlaneCoordinate(window, 0, sgct_core::SGCTProjectionPlane::LowerLeft);
             glm::vec2 vpLocalLowerLeft{ vpLocalLowerLeftA[0], vpLocalLowerLeftA[1] };
@@ -160,7 +160,7 @@ namespace viscom {
 
             // glm::vec2 relPosScale = 1.0f / glm::vec2(viewportQuadSize_[wId]);
             // glm::vec2 scaledRelPos = (glm::vec2(viewportScreen_[wId].position_) / glm::vec2(viewportScreen_[wId].size_)) * relPosScale;
-            
+
             glm::mat4 glbToLcMatrix = glm::mat4{ 1.0f };
             // correct local matrix:
             // xlocal = xglobal*totalScreenSize - viewportScreen_[wId].position_
@@ -236,18 +236,18 @@ POP_WARNINGS
 
     bool FrameworkInternal::IsMouseButtonPressed(int button) const noexcept
     {
-        return mousePressedState_[button];
+        return mousePressedState_[static_cast<std::size_t>(button)];
     }
 
     bool FrameworkInternal::IsKeyPressed(int key) const noexcept
     {
-        return keyPressedState_[key];
+        return keyPressedState_[static_cast<std::size_t>(key)];
     }
 
     void FrameworkInternal::BaseKeyboardCallback(int key, int scancode, int action, int mods)
     {
         if (!initialized_) return;
-        keyPressedState_[key] = (action == GLFW_RELEASE) ? false : true;
+        keyPressedState_[static_cast<std::size_t>(key)] = (action == GLFW_RELEASE) ? false : true;
 
         appNodeInternal_->KeyboardCallback(key, scancode, action, mods);
     }
@@ -262,7 +262,7 @@ POP_WARNINGS
     void FrameworkInternal::BaseMouseButtonCallback(int button, int action)
     {
         if (!initialized_) return;
-        mousePressedState_[button] = (action == GLFW_RELEASE) ? false : true;
+        mousePressedState_[static_cast<std::size_t>(button)] = (action == GLFW_RELEASE) ? false : true;
 
         appNodeInternal_->MouseButtonCallback(button, action);
     }
@@ -296,13 +296,16 @@ POP_WARNINGS
         auto internalID = reinterpret_cast<std::uint8_t*>(&splitID[0]);
         switch (static_cast<InternalTransferType>(internalID[0])) {
         case InternalTransferType::ResourceTransfer:
-            CreateSynchronizedResource(static_cast<ResourceType>(internalID[1]), receivedData, receivedLength);
+            CreateSynchronizedResource(static_cast<ResourceType>(internalID[1]), receivedData, static_cast<std::size_t>(receivedLength));
             break;
         case InternalTransferType::ResourceReleaseTransfer:
-            ReleaseSynchronizedResource(static_cast<ResourceType>(internalID[1]), std::string_view(reinterpret_cast<char*>(receivedData), receivedLength));
+            ReleaseSynchronizedResource(
+                static_cast<ResourceType>(internalID[1]),
+                std::string_view(reinterpret_cast<char*>(receivedData), static_cast<std::size_t>(receivedLength)));
             break;
         case InternalTransferType::ResourceRequest:
-            SendResourcesToNode(static_cast<ResourceType>(internalID[1]), receivedData, receivedLength, clientID);
+            SendResourcesToNode(static_cast<ResourceType>(internalID[1]), receivedData,
+                                static_cast<std::size_t>(receivedLength), clientID);
             break;
         default:
             LOG(WARNING) << "Unknown InternalTransferType: " << internalID[0];
@@ -523,14 +526,14 @@ POP_WARNINGS
             result.emplace_back(fboSize.x, fboSize.y, fboDesc);
             LOG(DBUG) << "Offscreen FBO VP Pos: " << 0.0f << ", " << 0.0f;
             LOG(DBUG) << "Offscreen FBO VP Size: " << fboSize.x << ", " << fboSize.y;
-            result.back().SetStandardViewport(0, 0, fboSize.x, fboSize.y);
+            result.back().SetStandardViewport(0, 0, static_cast<unsigned int>(fboSize.x), static_cast<unsigned int>(fboSize.y));
         }
         return result;
     }
 
     const FrameBuffer* FrameworkInternal::SelectOffscreenBuffer(const std::vector<FrameBuffer>& offscreenBuffers) const
     {
-        return &offscreenBuffers[engine_->getCurrentWindowPtr()->getId()];
+        return &offscreenBuffers[static_cast<std::size_t>(engine_->getCurrentWindowPtr()->getId())];
     }
 
     std::unique_ptr<FullscreenQuad> FrameworkInternal::CreateFullscreenQuad(const std::string& fragmentShader)
